Change returned type in value getter to int

+++ src/main/java/com/example/java/maven/ticketVendingMachine/Coin.java
@@ -19,7 +19,7 @@ public enum Coin {
         this.symbol = symbol;
     }
 
-    public float getValue() {
+    public int getValue() {
         return value;
     }
